Extract EKMGenerator into module
